move stuff not intended to be directly used to detail
